Updates for running on mre vm (Azure)
