+++ scripts/change-sdk-source.js
@@ -1,6 +1,7 @@
 var fs = require('fs');
 var shell = require('shelljs');
 var path = require('path');
+var forEachSample = require('./foreach-sample');
 
 const helpText = "Usage: change-sdk-source.js <sdk_source_type>\n\n\tWhere <sdk_source_type> is either: npm or sdk-source\n\n" +
     "\tEnsure that there is a 'sdk-path-config.json' containing { \"sdkPath\": \"<path_to_sdk_directory>\" }\n";
@@ -39,9 +40,6 @@ if (args.length === 0) {
     printError("Must supply an argument of either npm or sdk-source.");
 }
 
-const scriptPath = __dirname;
-const samplesPath = path.join(scriptPath, '..', 'samples');
-
 var config = JSON.parse(fs.readFileSync(`${__dirname}/sdk-path-config.json`, 'utf8'));
 if (!config || !config.sdkPath) {
     printError("Must provide a valid sdk-path-config.json with a valid sdkPath.");
@@ -50,11 +48,8 @@ else {
     config.sdkPath = path.resolve(__dirname, config.sdkPath);
 }
 
-fs.readdirSync(samplesPath).forEach((entry) => {
-    const sampleDir = path.join(samplesPath, entry);
-    if (fs.statSync(sampleDir).isDirectory()) {
-        updateSdkTarget(sampleDir, config.sdkPath);
-    }
-})
+forEachSample((sampleDir) => {
+    updateSdkTarget(sampleDir, config.sdkPath);
+});
 
 process.exit();
